Implement DI and DIUN
This implements the Device Initalize and Device Initialize over
Untrusted Networks (DIUN) protocols.
The DIUN protocol spec is updated to the current version of the draft
protocol, and all the key attestation bits are removed for now.
Key attestation will be brought back in the future, after more
consideration.

There are currently no integration tests for these protocols, those will
be added next.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>