fix: bring PublicKey back to certificates
In PR #48, we changed the PublicKey structure to DER encoded public
keys, because that's how the specification reads.
Unfortunately, that is not the way it was intended: intended was an
actual DER-encoded X509 certificate (X509), or an x5chain (COSEX509).
This will be clarified in an upcoming erratum to the specification.
This patch brings it back to the correct structures.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>