feat: data format rework
Ensure that all data formats are (de)serialized in the correct manner,
and that we compute accurate digests.
In the FDO spec, some data structures (notably the Ownership Voucher and
its header) contain data that needs to get digests computed as part of
the protocol.

Unfortunately, due to the fact that these structures are sent "fat",
i.e. they're fully inline, rather than wrapper by a bstr, we were
required to re-serialize them, in order to compute a digest.
This results in digests not matching between different implementations,
because of the fact that the same thing can be represented in multiple
ways in CBOR.

To prevent this, this patch implements a custom CBOR parser, which is
very limited: it only parses structures that are an Array (of indefinite
length).
It also is very restrictive in what it allows (no indefinite length at
all, no special values, etc), but those assumptions hold for FDO and its
data structures.
This parser then keeps track of the exact bytes that were read, and as a
consequence it allows us to compute hashes over exactly the same data as
we received.

This patch also updates the rest of the codebase to use the new parser.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>