feat: use ciborium for serializing
The specification does not require CBOR canonical encoding, but some
other implementations do, due to the amount of nested data structures
that need to get hashes.
In order to make sure we generate contents they can correctly hash, this
changes us to use ciborium for serializing, since that generates
canonical CBOR (for the parts we care about).
This means we now have two CBOR libraries, and we might want to move to
ciborium fully further down the line, but at this moment using it for
just serializing is the easy-fix that gets us compatible with other
implementations.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>