feat: add non-interoperable KDF
This adds the ability to add a non-interoperable KDF implementation.
This should never be used in production environments, as it will be not
interoperable with other implementations, or this implementation with
the interoperable KDF.

When the non-interoperable KDF is enabled, the client and server will
send a header to indicate this, so it becomes possible for the other
side to dynamically fall back to the non-interoperable KDF when needed.
This is work for a future patch, but this ensures that such a patch will
have the required information.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>