feat: implement semi-TPM Device Credential
This implements a system where the keys for the device credential are
created by the TPM (never leaving the TPM), but the references to the
keys, and rendezvous info, etc, are stored on the filesystem.
This protects against having the actual private keys stolen when the
device credential file is extracted from a disk.

The current implementation only supports the ECDSA ciphers mentioned by
the FDO specification.

The current test suite does not set up a software TPM to execute the TPM
tests during CI, but it will execute the TPM tests if a TPM exists on
the system.

Fixes: #200
Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>