chore: update example rendezvous-server.yml
Update rendezvous-server.yml to include 'max_wait_seconds:' and updated storage path used in the
AIO. This prevents a panic when trying to run the server without 'max_wait_seconds'.

This also removes the 'trusted_device_keys_path: /path/to/keys/device_ca_cert.pem', as it
doesn't seem to be needed (not included in the config when running the fdo-aio).

Please confirm before merging.

Signed-off-by: Paul Whalen <pwhalen@redhat.com>