feat: add an option to reboot once onboarding is completed
This adds a new serviceinfo api module 'reboot' which allows
to specify if the system should be rebooted after onboarding
is completed.

This means that no client errors can be sent back since onboarding
is already considered to be a success. This reboot action is
always handled last, so any other onboarding action may fail
and will be handled according to the protocol. Only if onboarding
is successfully handled the reboot action will take place.

This reboot action is optional and can be configured in the
'serviceinfo-api-server.yml'.

Signed-off-by: Irene Diez <idiez@redhat.com>