fix(serviceinfo): no authentication is needed when omitting `service_info_auth_token`
This patch supports no authentication for service info by using Option
for `service_info_auth_token`.

Currentl docs mentions `None` if no authentication is needed but
`serde_yaml` library does not support it and so
`service_info_auth_token` needs to support the `Option()`.

Alternatively it is possible to support `None` or `""` (empty string)
as a special case to disable authorization. Any suggestion is welcome.