feat(admin-tool): configure X509 certificate's expiration date
Currently with the `generate-key-and-cert` option of the
`fdo-admin-tool` cli we can configure the 'Country' and
'Organization' name of the generated X509 certificate, and we
set by default a 365-day validity period.

This allows to set an specific number of days for the certificate's
validity period with the new `--validity-ends` option.

Signed-off-by: Irene Diez <idiez@redhat.com>