chore: replace sha-crypt with openssl process calls
This replaces the sha-crypt crate and uses straight  `openssl` calls in the
code. This also changes the hashing alogirithm to sha256, and removes some
checks in the tests that are no longer relevant. Calling `openssl` directly
makes it use a different salt for each hash so we cannot test that the
passwords are what we expect anymore.

Signed-off-by: Irene Diez <idiez@redhat.com>