fix: change TCTI type from Tabrmd to Device (Kernel RM)
We were using the Tabrmd as the way to communicate with the TPM, however we
should be using the Kernel resource manager instead. This commit changes that
and we default to using `/dev/tpmrm0` if no other named config is passed via
environment variables (TPM2TOOLS_TCTI, TCTI, TEST_TCTI).

Signed-off-by: Irene Diez <idiez@redhat.com>