fix: remove using the TPM by default in aio
The `aio` testing tool for running all the servers in the FDO infrastructure
enabled using TPM support by default in the manufacturing_server, which
caused issues when the tests were run in hosts that did not have a TPM or it
was misconfigured.

To avoid further bugs this commit turns the TPM option off by default.

Signed-off-by: Irene Diez <idiez@redhat.com>