feat(fsim-csr): implement the  CA certs enrollment operation on device side
The device sends the active modules to the owner's onboarding server.
If the CSR module is among the active modules, the owner's onboarding server
 will send back the 'fdo.csr.cacerts-res' message to the device encoded
as described in the specification[1]. Te device will decode the message,
save the file in the system's trusted anchors and update the certificates
trusted by the system.

[1] https://github.com/fido-alliance/fdo-sim/blob/main/fsim-repository/fdo.csr.md

Signed-off-by: Miguel Martín <mmartinv@redhat.com>