fix!: hashing of device cert chain in ownership voucher
Implement hashing behavior for device certificate chain in ownership
voucher as dictated by the specification

The specification states that “OVDevCertChainHash is the Hash of the
concatenation of the contents of each byte string in
OwnershipVoucher.OVDevCertChain, in the presented order."

The current behavior is to serialize the X5CHAIN to CBOR (i.e.
`[ bstr[cert1] ... bstr[certN] ]` instead of `cert1 || ... || certN`).

BREAKING CHANGE: The device certificate chain hash in ownership voucher
is now calculated from the concatenation of certificate ASN.1 DER
contents rather than the CBOR serialization of an array of byte strings.

Signed-off-by: Ben Krieger <ben.krieger@intel.com>