feat(manufacturing-server): implement an export OVs endpoint
Just serving an archive with all the OVs the manufacturer knows about.
It'd be handy to just give this away to whoever needs these credentials
and/or create a nice UI where you click a button to have them all.
The post-MVP, with a UI, would be to just have a UI that is able to list
all the device credentials, let you select which one you want, download
them in an archive, profit. Not there yet.

Signed-off-by: Antonio Murdaca <amurdaca@redhat.com>