fix: use centos-stream-9 target instead of epel-9
Use the centos-stream-9 target as the epel-9 repos
don't seem to be up to date.

Remove the preparation steps enabling the copr repo
and installing the copr packages because it's not needed
and it's causing problems: Packit is supposed to enable the
copr repo and install all the packages contained in
that repo.

Signed-off-by: Miguel Martín <mmartinv@redhat.com>