fix(make-vendored-tarfile.sh): exclude idna tests with unicode points
I imagine we could get rid of all the tests/ directories in every create
but for now, this addresses rpmlint in centos/rhel.
See https://url.corp.redhat.com/a907bc8
We usually waived this kind of errors in the past but this is a good
time and easy patch to fix it.

Signed-off-by: Antonio Murdaca <amurdaca@redhat.com>