chore: bump aws-nitro-enclaves-cose to 0.5.2
Use the latest released version of aws-nitro-enclaves-cose (v0.5.2) in
`data-formats` and `http-wrapper`.

Replace the dependency above in the workspace so it contains the fix
that allow the use of CoseError in a multi-threaded environment:

https://github.com/awslabs/aws-nitro-enclaves-cose/pull/115

Once a new version of aws-nitro-enclaves-cose is released we can
remove the replacement at workspace level an bump the version again
in `data-formats` and `http-wrapper`.

Update the code to reflect the latest API changes made
by aws-nitro-enclaves-cose.

Signed-off-by: Miguel Martín <mmartinv@redhat.com>